refactor(api): restructure API handlers for improved clarity and organization
- Moved GET request handling logic into dedicated functions within a new getRequests module.
- Updated API routes to utilize the new handler functions, enhancing readability and maintainability.
- Removed redundant code from the main API file, adhering to the single-responsibility principle.
- Added new tests for GET requests to ensure functionality and coverage.
