+++ packages/client-direct/__test__/direct.test.ts
@@ -100,110 +100,9 @@ describe("DirectClient", () => {
         client.stop();
     });
 
-    describe("GET /", () => {
-        it("should return welcome message", async () => {
-            const response = await request(client.app).get("/");
-            expect(response.status).toBe(200);
-            expect(response.text).toBe("Welcome, this is the REST API!");
-        });
-    });
 
-    describe("GET /hello", () => {
-        it("should return hello world message", async () => {
-            const response = await request(client.app).get("/hello");
-            expect(response.status).toBe(200);
-            expect(response.body).toEqual({ message: "Hello World!" });
-        });
-    });
 
-    describe("GET /agents", () => {
-        it("should return list of agents", async () => {
-            const response = await request(client.app).get("/agents");
-            expect(response.status).toBe(200);
-            expect(response.body).toEqual({
-                agents: [
-                    {
-                        id: mockAgentRuntime.agentId,
-                        name: "Test Agent",
-                        clients: ["discord"],
-                    },
-                ],
-            });
-        });
 
-        it("should return empty list when no agents exist", async () => {
-            // @ts-ignore: even though it's private, we can clear it for testing
-            client.agents.clear();
-            const response = await request(client.app).get("/agents");
-            expect(response.status).toBe(200);
-            expect(response.body).toEqual({ agents: [] });
-        });
-    });
-
-    describe("Constructor and Basic Setup", () => {
-        it("should initialize with express app and middleware", () => {
-            expect(client.app).toBeDefined();
-            expect(client.app._router).toBeDefined();
-        });
-    });
-
-    describe("GET /agents/:agentId/channels", () => {
-        it("should return agent channels when valid token is provided", async () => {
-            const mockGuilds = [
-                { id: "guild1", name: "Guild 1" },
-                { id: "guild2", name: "Guild 2" },
-            ];
-
-            // Mock Discord.js REST
-            const mockGet = vi.fn().mockResolvedValue(mockGuilds);
-            vi.mocked(REST).mockImplementation(
-                () =>
-                    ({
-                        setToken: () => ({ get: mockGet }),
-                    }) as any
-            );
-
-            const response = await request(client.app).get(
-                `/agents/${mockAgentRuntime.agentId}/channels`
-            );
-
-            expect(response.status).toBe(200);
-            expect(response.body).toEqual({
-                id: mockAgentRuntime.agentId,
-                guilds: mockGuilds,
-                serverCount: 2,
-            });
-        });
-
-        it("should return 404 when agent is not found", async () => {
-            const response = await request(client.app).get(
-                "/agents/non-existent-agent/channels"
-            );
-            expect(response.status).toBe(404);
-            expect(response.body).toEqual({ error: "Agent not found" });
-        });
-
-        it("should handle Discord API errors", async () => {
-            vi.mocked(REST).mockImplementation(
-                () =>
-                    ({
-                        setToken: () => ({
-                            get: vi
-                                .fn()
-                                .mockRejectedValue(
-                                    new Error("Discord API Error")
-                                ),
-                        }),
-                    }) as any
-            );
-
-            const response = await request(client.app).get(
-                `/agents/${mockAgentRuntime.agentId}/channels`
-            );
-            expect(response.status).toBe(500);
-            expect(response.body.error).toBe("Error processing channels");
-        });
-    });
 
     describe("Agent Registration", () => {
         it("should register and unregister agents", () => {
