+++ packages/client-direct/src/handlers/index.ts
@@ -1,0 +1,1 @@
+export * as getRequests from "./getRequests";
